Jour 6 - Finition du site

+++ www/server/views/showHome.php
@@ -19,38 +19,66 @@ Description : La vue de la page d'accueil contenant toutes les annonces valides
         </div>
         <div class='row justify-content-center'>
             <div class='col-md-10'>
-                <form class='form-row'>
-                    <input type='text' class='form-control col-8' name='searchContent' placeholder='Mettez votre recherche ici'>
-                    <input type='submit' class='btn btn-outline-primary col my-btn' value='Rechercher'>
-                    <?php if (isset($_SESSION['email'])) : ?>
-                        <a href='../controllers/createAd.php' class='btn btn-outline-success col my-btn'>Ajouter</a>
-                    <?php endif; ?>
+                <form method='POST' class='m-1'>
+                    <div class='form-row'>
+                        <div class='form-group col-8'>
+                            <input type='text' class='form-control' name='searchContent' placeholder="Chercher par canton, ville, code postal, score, titre et description d'une annonce">
+                        </div>
+                        <div class='form-group col'>
+                            <input type='submit' class='form-control btn btn-outline-primary my-btn' value='Rechercher' name='search'>
+                        </div>
+                        <?php if (isset($_SESSION['email'])) : ?>
+                            <div class='form-group col'>
+                                <a href='../controllers/createAd.php' class='form-control btn btn-outline-success my-btn '>Ajouter</a>
+                            </div>
+                        <?php endif; ?>
+                    </div>
+                    <div class='form-row'>
+                        <div class='form-group'>
+                            <input type='checkbox' id='bio' name='searchOrganic'>
+                            <label for='bio'>Bio</label>
+                        </div>
+                    </div>
                 </form>
             </div>
         </div>
 
         <div class='row justify-content-center'>
-            <?php foreach ($ads as $ad) :
-                $u = UserManager::GetUserByEmail($ad->userEmail);
-                ?>
-                <div class='media border rounded col-md-10 my-ad'>
-                    <div class='media-body'>
-                        <h4><?= $ad->title ?></h4>
-                        <p class='text-justify'><?= $ad->description ?></p>
-                        <input type='radio' id='bio' <?= ($ad->organic == 1 ? 'checked' : '') ?>>
-                        <label for='bio'>Produit bio</label>
-                        <p><?= 'Vente au ' . $u->streetAndNumber . ', ' . $u->postCode . ' ' . $u->canton ?></p>
-                        <p><?= 'Posté le ' . date_format(date_create($ad->creationDate), 'd M Y \à H:i:s') ?></p>
-                        <div class='row justify-content-end'>
-                            <a class='btn btn-outline-primary my-btn' href='../controllers/adDetails.php?idAd=<?= $ad->idAdvertisement ?>'>Détails</a>
-                            <?php if (isset($_SESSION['email']) && $ad->userEmail == $_SESSION['email']) : ?>
-                                <a class='btn btn-outline-warning my-btn' href='../controllers/updateAd.php?idAd=<?= $ad->idAdvertisement ?>'>Modifier</a>
-                                <a class='btn btn-outline-danger my-btn' href='../controllers/deleteAd.php?idAd=<?= $ad->idAdvertisement ?>'>Supprimer</a>
+            <?php if (count($ads) > 0) :
+                foreach ($ads as $ad) :
+                    $u = UserManager::GetUserByEmail($ad->userEmail);
+                    $score = RatingManager::GetScoreOfAnAd($ad->idAdvertisement);
+                    ?>
+                    <div class='media border rounded col-md-10 my-ad'>
+                        <div class='media-body'>
+                            <h4><?= $ad->title ?></h4>
+                            <p class='text-justify'><?= $ad->description ?></p>
+                            <?php if ($ad->organic == 1) : ?>
+                                <span class="fa fa-pagelines"></span>
+                                <label>Produit bio</label>
+                            <?php else : ?>
+                                <label>Produit non bio</label>
                             <?php endif; ?>
+                            <p><?= 'Vente au ' . $u->streetAndNumber . ', ' . $u->postCode . ' ' . $u->canton ?></p>
+                            <p><?= 'Posté le ' . date_format(date_create($ad->creationDate), 'd M Y \à H:i:s') ?></p>
+                            <?php if ($score == 0) : ?>
+                                <p>Il n'y a aucune évalution de cette annonce</p>
+                            <?php else : ?>
+                                <p>Évalutation de l'annonce par les autres utilisateurs : <?= $score ?>/5</p>
+                            <?php endif; ?>
+                            <div class='row justify-content-end'>
+                                <a class='btn btn-outline-primary my-btn' href='../controllers/adDetails.php?idAd=<?= $ad->idAdvertisement ?>'>Détails</a>
+                                <?php if (isset($_SESSION['email']) && $ad->userEmail == $_SESSION['email']) : ?>
+                                    <a class='btn btn-outline-warning my-btn' href='../controllers/updateAd.php?idAd=<?= $ad->idAdvertisement ?>'>Modifier</a>
+                                    <a class='btn btn-outline-danger my-btn' href='../controllers/deleteAd.php?idAd=<?= $ad->idAdvertisement ?>'>Supprimer</a>
+                                <?php endif; ?>
+                            </div>
                         </div>
                     </div>
-                </div>
-            <?php endforeach; ?>
+                <?php endforeach; ?>
+            <?php else : ?>
+                <h2>Aucune annonce avec cette recherche</h2>
+            <?php endif; ?>
         </div>
         <div class='row'>
             <?php include_once '../inc/footer.php'; ?>
